# API_M3_V5/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API_M3",
    "description": "XTREME API",
    "version": "v5"
  },
  "paths": {
...
    "/Client/ClientById": {
      "get": {
        "tags": [
          "Client"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "cli_id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/Client/ClientByChar": {
      "get": {
        "tags": [
          "Client"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "type": "string"
          },
          {
            "in": "query",
            "name": "value",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/Client/NewClient": {
      "post": {
        "tags": [
          "Client"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Client_aux"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/Client/UpClient": {
      "put": {
        "tags": [
          "Client"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Client_aux"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Client_aux": {
      "type": "object",
      "properties": {
        "clientId": {
          "format": "int32",
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "personId": {
          "format": "int32",
          "type": "integer"
        },
        "entityId": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "nif": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "format": "int32",
          "type": "integer"
        },
        "addressline": {
          "type": "string"
        },
        "zipcode": {
          "format": "int32",
          "type": "integer"
        },
        "locationId": {
          "format": "int32",
          "type": "integer"
        },
        "city": {
          "type": "string"
        },
        "district": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      }
    },
...
  }
}